# Description
Reduced the evaluation time on running evaulator.py by using multiprocessing

# Performance Report *
| Agent Name | Time elapsed without using multiprocessing (in second) |  Time elapsed using multiprocessing (in second) |
--------| ----------- | --------- |
Word Frequency Similarity  | 94.93 | 44.15 |
tfidf | 682.89 | 757.59 |
Ngram | 711.2 | 115.71 |
DLD | 1021.62 | 526.53 |

\* The time taken also depend on hardware specification and may vary on different computers

Complete Report can be found [here](https://docs.google.com/document/d/1bnuM9IzDl82RkhYJrSlrxxmnBcEI1JjawfX31wQTxBc/edit?usp=sharing)

Closes #64