@GMishx 2020-12-29
The changes seems to break `tqdm` progress bar. Can you please try to fix it? Otherwise it can be removed.